rvcapi_inference/requirements.txt
numba==0.56.4
numpy==1.23.5
scipy==1.9.3
librosa==0.9.2
llvmlite==0.39.0
fairseq==0.12.2
faiss-cpu==1.7.0; sys_platform == "darwin"
faiss-cpu==1.7.2; sys_platform != "darwin"
gradio
Cython
future>=0.18.3
pydub>=0.25.1
soundfile>=0.12.1
ffmpeg-python>=0.2.0
tensorboardX
functorch>=2.0.0
Jinja2>=3.1.2
json5>=0.9.11
Markdown
matplotlib>=3.7.1
matplotlib-inline>=0.1.6
praat-parselmouth>=0.4.3
Pillow>=9.1.1
pyworld>=0.3.2
resampy>=0.4.2
scikit-learn>=1.2.2
starlette>=0.26.1
tensorboard
tensorboard-data-server
tensorboard-plugin-wit
torchgen>=0.0.1
tqdm>=4.65.0
tornado>=6.2
Werkzeug>=2.2.3
uc-micro-py>=1.0.1
sympy>=1.11.1
tabulate>=0.9.0
PyYAML>=6.0
pyasn1>=0.4.8
pyasn1-modules>=0.2.8
fsspec>=2023.3.0
absl-py>=1.4.0
audioread
uvicorn>=0.21.1
colorama==0.4.2
mongoengine==0.27.0
setuptools==57.5.0
fastapi